[RBS-1084][DS] modified and moved types, updated relevant files
